+++ src/openepd/model/lcia.py
@@ -42,14 +42,28 @@ class EolScenario(BaseOpenEpdSchema):
         description="The weigting of this scenario used in the C1 .. C4 dataset. For example, the overall C1 shoudl be "
         "equal to weighted sum of C1 from all the scenarios, weighted by likelihood.",
         example=0.33,
+        default=None,
     )
-    C1: Measurement | None = pyd.Field(description="Deconstruction and Demolition under this scenario")
-    C2: Measurement | None = pyd.Field(description="Transport to waste processing or disposal under this scenario.")
-    C3: Measurement | None = pyd.Field(description="Waste Processing under this scenario")
-    C4: Measurement | None = pyd.Field(description="Disposal under this scenario")
+    C1: Measurement | None = pyd.Field(
+        description="Deconstruction and Demolition under this scenario",
+        default=None,
+    )
+    C2: Measurement | None = pyd.Field(
+        description="Transport to waste processing or disposal under this scenario.",
+        default=None,
+    )
+    C3: Measurement | None = pyd.Field(
+        description="Waste Processing under this scenario",
+        default=None,
+    )
+    C4: Measurement | None = pyd.Field(
+        description="Disposal under this scenario",
+        default=None,
+    )
     D: Measurement | None = pyd.Field(
         description="Potential net benefits from reuse, recycling, and/or energy recovery beyond "
-        "the system boundary under this scenario"
+        "the system boundary under this scenario",
+        default=None,
     )
 
 
@@ -61,50 +75,128 @@ class ScopeSet(BaseOpenEpdSchema):
     The 'unit' field must be consistent across all scopes in a single scopeset.
     """
 
-    A1A2A3: Measurement | None = pyd.Field(description="Sum of A1..A3")
-    A1: Measurement | None = pyd.Field(description="Raw Material Supply")
-    A2: Measurement | None = pyd.Field(description="Transport to Manufacturing")
-    A3: Measurement | None = pyd.Field(description="Manufacturing")
-    A4: Measurement | None = pyd.Field(description="Transport to Construction")
-    A5: Measurement | None = pyd.Field(description="Construction")
-    B1: Measurement | None = pyd.Field(description="Use impacts over Reference Service Life (Predicted)")
+    A1A2A3: Measurement | None = pyd.Field(
+        description="Sum of A1..A3",
+        default=None,
+    )
+    A1: Measurement | None = pyd.Field(
+        description="Raw Material Supply",
+        default=None,
+    )
+    A2: Measurement | None = pyd.Field(
+        description="Transport to Manufacturing",
+        default=None,
+    )
+    A3: Measurement | None = pyd.Field(
+        description="Manufacturing",
+        default=None,
+    )
+    A4: Measurement | None = pyd.Field(
+        description="Transport to Construction",
+        default=None,
+    )
+    A5: Measurement | None = pyd.Field(
+        description="Construction",
+        default=None,
+    )
+    B1: Measurement | None = pyd.Field(
+        description="Use impacts over Reference Service Life (Predicted)",
+        default=None,
+    )
     B1_years: float | None = pyd.Field(
         gt=0,
         lt=11,
         description="Timeframe over which B1 is evaluated, in years. "
         "For example, an impact of 1.23 kgCO2e per year for ten years could be B1=1.23kgCO2e, "
         "B1_years=1.0 or  B1=12.3kgCO2e, B1_years=10.0",
+        default=None,
     )
-    B2: Measurement | None = pyd.Field(description="Predicted Maintenance Impacts over Reference Service Life")
+    B2: Measurement | None = pyd.Field(
+        description="Predicted Maintenance Impacts over Reference Service Life",
+        default=None,
+    )
     B2_years: float | None = pyd.Field(
-        gt=0, lt=11, description="Predicted Maintenance Impacts over Reference Service Life"
+        gt=0,
+        lt=11,
+        description="Predicted Maintenance Impacts over Reference Service Life",
+        default=None,
     )
-    B3: Measurement | None = pyd.Field(description="Predicted Repair impacts over Reference Service Life")
-    B3_years: float | None = pyd.Field(gt=0, lt=11, description="Timeframe over which B3 is evaluated, in years")
+    B3: Measurement | None = pyd.Field(
+        description="Predicted Repair impacts over Reference Service Life",
+        default=None,
+    )
+    B3_years: float | None = pyd.Field(
+        gt=0,
+        lt=11,
+        description="Timeframe over which B3 is evaluated, in years",
+        default=None,
+    )
     B4: Measurement | None = pyd.Field(
         description="Predicted Replacement Impacts over the Building lifetime "
-        "('Estimated Construction Works lifespan') specified in the PCR."
+        "('Estimated Construction Works lifespan') specified in the PCR.",
+        default=None,
     )
-    B4_years: float | None = pyd.Field(gt=0, lt=11, description="Timeframe over which B4 is evaluated, in years")
+    B4_years: float | None = pyd.Field(
+        gt=0,
+        lt=11,
+        description="Timeframe over which B4 is evaluated, in years",
+        default=None,
+    )
     B5: Measurement | None = pyd.Field(
         description="Predicted Refurbishment Impacts over the Building lifetime "
-        "('Estimated Construction Works lifespan') specified in the PCR."
+        "('Estimated Construction Works lifespan') specified in the PCR.",
+        default=None,
     )
-    B5_years: float | None = pyd.Field(gt=0, lt=11, description="Timeframe over which B5 is evaluated, in years")
-    B6: Measurement | None = pyd.Field(description="Predicted Impacts related to Operational Energy Use")
-    B6_years: float | None = pyd.Field(gt=0, lt=11, description="Timeframe over which B6 is evaluated, in years")
-    B7: Measurement | None = pyd.Field(description="Predicted Impacts related to Operational Water Use")
-    B7_years: float | None = pyd.Field(gt=0, lt=11, description="Timeframe over which B7 is evaluated, in years")
+    B5_years: float | None = pyd.Field(
+        gt=0,
+        lt=11,
+        description="Timeframe over which B5 is evaluated, in years",
+        default=None,
+    )
+    B6: Measurement | None = pyd.Field(
+        description="Predicted Impacts related to Operational Energy Use",
+        default=None,
+    )
+    B6_years: float | None = pyd.Field(
+        gt=0,
+        lt=11,
+        description="Timeframe over which B6 is evaluated, in years",
+        default=None,
+    )
+    B7: Measurement | None = pyd.Field(
+        description="Predicted Impacts related to Operational Water Use",
+        default=None,
+    )
+    B7_years: float | None = pyd.Field(
+        gt=0,
+        lt=11,
+        description="Timeframe over which B7 is evaluated, in years",
+        default=None,
+    )
     C_scenarios: list[EolScenario] | None = pyd.Field(
         description="A list of possible end-of-life scenarios, "
-        "for use in analyses where the end-of-life can be predicted."
+        "for use in analyses where the end-of-life can be predicted.",
+        default=None,
     )
-    C1: Measurement | None = pyd.Field(description="Deconstruction and Demolition")
-    C2: Measurement | None = pyd.Field(description="Transport to waste processing or disposal.")
-    C3: Measurement | None = pyd.Field(description="Waste Processing")
-    C4: Measurement | None = pyd.Field(description="Disposal")
+    C1: Measurement | None = pyd.Field(
+        description="Deconstruction and Demolition",
+        default=None,
+    )
+    C2: Measurement | None = pyd.Field(
+        description="Transport to waste processing or disposal.",
+        default=None,
+    )
+    C3: Measurement | None = pyd.Field(
+        description="Waste Processing",
+        default=None,
+    )
+    C4: Measurement | None = pyd.Field(
+        description="Disposal",
+        default=None,
+    )
     D: Measurement | None = pyd.Field(
-        description="Potential net benefits from reuse, recycling, and/or energy recovery beyond the system boundary."
+        default=None,
+        description="Potential net benefits from reuse, recycling, and/or energy recovery beyond the system boundary.",
     )
 
 
@@ -112,26 +204,28 @@ class ImpactSet(BaseOpenEpdSchema):
     """A set of impacts, such as GWP, ODP, AP, EP, POCP, EP-marine, EP-terrestrial, EP-freshwater, etc."""
 
     gwp: ScopeSet | None = pyd.Field(
+        default=None,
         description="GWP100, calculated per IPCC guidelines.  If any CO2 removals are "
         "part of this figure, the gwp-fossil, gwp-bioganic, gwp-luluc, an "
         "gwp-nonCO2 fields are required, as is "
-        "kg_C_biogenic_per_declared_unit."
+        "kg_C_biogenic_per_declared_unit.",
     )
-    odp: ScopeSet | None = pyd.Field(description="Ozone Depletion Potential")
-    ap: ScopeSet | None = pyd.Field(description="Acidification Potential")
+    odp: ScopeSet | None = pyd.Field(default=None, description="Ozone Depletion Potential")
+    ap: ScopeSet | None = pyd.Field(default=None, description="Acidification Potential")
     ep: ScopeSet | None = pyd.Field(
-        description="Eutrophication Potential in Marine Ecosystems. Has the same meaning as ep-marine."
+        default=None, description="Eutrophication Potential in Marine Ecosystems. Has the same meaning as ep-marine."
     )
-    pocp: ScopeSet | None = pyd.Field(description="Photochemical Smog (Ozone) creation potential")
-    ep_marine: ScopeSet | None = pyd.Field(alias="ep-marine", description="Has the same meaning as 'ep'")
+    pocp: ScopeSet | None = pyd.Field(default=None, description="Photochemical Smog (Ozone) creation potential")
+    ep_marine: ScopeSet | None = pyd.Field(alias="ep-marine", default=None, description="Has the same meaning as 'ep'")
     ep_fresh: ScopeSet | None = pyd.Field(
-        alias="ep-fresh", description="Eutrophication Potential in Freshwater Ecosystems"
+        alias="ep-fresh", default=None, description="Eutrophication Potential in Freshwater Ecosystems"
     )
     ep_terr: ScopeSet | None = pyd.Field(
-        alias="ep-terr", description="Eutrophication Potential in Terrestrial Ecosystems"
+        alias="ep-terr", default=None, description="Eutrophication Potential in Terrestrial Ecosystems"
     )
     gwp_biogenic: ScopeSet | None = pyd.Field(
         alias="gwp-biogenic",
+        default=None,
         description="Net GWP from removals of atmospheric CO2 into biomass and  emissions of CO2 from biomass sources. "
         "To be counted as negative, CO2 removals must be closely linked to production in "
         "operational control (i.e. no purchased offsets), time (within three years of production) and "
@@ -140,6 +234,7 @@ class ImpactSet(BaseOpenEpdSchema):
     )
     gwp_luluc: ScopeSet | None = pyd.Field(
         alias="gwp-luluc",
+        default=None,
         description="Climate change effects related to land use and land use change, for example biogenic carbon "
         "exchanges resulting from deforestation or other soil activities (including soil carbon "
         "emissions). All related emissions for native forests are included under this category. "
@@ -147,10 +242,12 @@ class ImpactSet(BaseOpenEpdSchema):
     )
     gwp_nonCO2: ScopeSet | None = pyd.Field(
         alias="gwp-nonCO2",
+        default=None,
         description="GWP from non-CO2, non-fossil sources, such as livestock-sourced CH4 and agricultural N2O.",
     )
     gwp_fossil: ScopeSet | None = pyd.Field(
         alias="gwp-fossil",
+        default=None,
         description="Climate change effects due to greenhouse gas emissions originating from the oxidation or "
         "reduction of fossil fuels or materials containing fossil carbon. [Source: EN15804]",
     )
@@ -230,47 +327,121 @@ class ResourceUseSet(BaseOpenEpdSchema):
     RPRec: ScopeSet | None = pyd.Field(
         description="Renewable primary resources used as energy carrier (fuel). "
         "First use bio-based materials used as an energy source. Hydropower, solar and wind power used "
-        "in the technosphere are also included in this indicator"
+        "in the technosphere are also included in this indicator",
+        default=None,
     )
     RPRm: ScopeSet | None = pyd.Field(
         description="Renewable primary resources with energy content used as material. "
-        "First use biobased materials used as materials (e.g. wood, hemp, etc.)."
+        "First use biobased materials used as materials (e.g. wood, hemp, etc.).",
+        default=None,
     )
-    rpre: ScopeSet | None = pyd.Field(description="Renewable primary energy resources as energy")
-    nrpre: ScopeSet | None = pyd.Field(description="Non-renewable primary resources as energy (fuel)")
-    nrprm: ScopeSet | None = pyd.Field(description="Non-renewable primary resources as material")
-    fw: ScopeSet | None = pyd.Field(description="Use of net fresh water")
-    sm: ScopeSet | None = pyd.Field(description="Use of secondary materials")
-    rsf: ScopeSet | None = pyd.Field(description="Use of renewable secondary materials")
-    nrsf: ScopeSet | None = pyd.Field(description="Use of non-renewable secondary fuels")
-    re: ScopeSet | None = pyd.Field(description="Renewable energy resources")
+    rpre: ScopeSet | None = pyd.Field(
+        description="Renewable primary energy resources as energy",
+        default=None,
+    )
+    nrpre: ScopeSet | None = pyd.Field(
+        description="Non-renewable primary resources as energy (fuel)",
+        default=None,
+    )
+    nrprm: ScopeSet | None = pyd.Field(
+        description="Non-renewable primary resources as material",
+        default=None,
+    )
+    fw: ScopeSet | None = pyd.Field(
+        description="Use of net fresh water",
+        default=None,
+    )
+    sm: ScopeSet | None = pyd.Field(
+        description="Use of secondary materials",
+        default=None,
+    )
+    rsf: ScopeSet | None = pyd.Field(
+        description="Use of renewable secondary materials",
+        default=None,
+    )
+    nrsf: ScopeSet | None = pyd.Field(
+        description="Use of non-renewable secondary fuels",
+        default=None,
+    )
+    re: ScopeSet | None = pyd.Field(
+        description="Renewable energy resources",
+        default=None,
+    )
     pere: ScopeSet | None = pyd.Field(
-        description="Use of renewable primary energy excluding renewable primary energy resources used as raw materials"
+        description="Use of renewable primary energy excluding renewable primary energy resources used as raw materials",
+        default=None,
     )
-    perm: ScopeSet | None = pyd.Field(description="Use of renewable primary energy resources used as raw materials")
-    pert: ScopeSet | None = pyd.Field(description="Total use of renewable primary energy resources")
+    perm: ScopeSet | None = pyd.Field(
+        description="Use of renewable primary energy resources used as raw materials",
+        default=None,
+    )
+    pert: ScopeSet | None = pyd.Field(
+        description="Total use of renewable primary energy resources",
+        default=None,
+    )
     penre: ScopeSet | None = pyd.Field(
         description="Use of non-renewable primary energy excluding "
-        "non-renewable primary energy resources used as raw materials"
+        "non-renewable primary energy resources used as raw materials",
+        default=None,
     )
     penrm: ScopeSet | None = pyd.Field(
-        description="Use of non-renewable primary energy resources used as raw materials"
+        description="Use of non-renewable primary energy resources used as raw materials",
+        default=None,
     )
-    penrt: ScopeSet | None = pyd.Field(description="Total use of non-renewable primary energy resources")
+    penrt: ScopeSet | None = pyd.Field(
+        description="Total use of non-renewable primary energy resources",
+        default=None,
+    )
 
 
 class OutputFlowSet(BaseOpenEpdSchema):
     """A set of output flows, such as waste, emissions, etc."""
 
-    twd: ScopeSet | None = pyd.Field(description="Total waste disposed")
-    hwd: ScopeSet | None = pyd.Field(description="Hazardous waste disposed")
-    nhwd: ScopeSet | None = pyd.Field(description="Non-hazardous waste disposed")
-    rwd: ScopeSet | None = pyd.Field(description="Radioactive waste disposed")
-    hlrw: ScopeSet | None = pyd.Field(description="High level radioactive waste disposed")
-    illrw: ScopeSet | None = pyd.Field(description="Intermediate level radioactive waste disposed")
-    cru: ScopeSet | None = pyd.Field(description="Components for re-use")
-    mr: ScopeSet | None = pyd.Field(description="Recycled materials")
-    mfr: ScopeSet | None = pyd.Field(description="Materials for recycling")
-    mer: ScopeSet | None = pyd.Field(description="Materials for energy recovery")
-    ee: ScopeSet | None = pyd.Field(description="Exported energy")
-    eh: ScopeSet | None = pyd.Field(description="Exported heat")
+    twd: ScopeSet | None = pyd.Field(
+        description="Total waste disposed",
+        default=None,
+    )
+    hwd: ScopeSet | None = pyd.Field(
+        description="Hazardous waste disposed",
+        default=None,
+    )
+    nhwd: ScopeSet | None = pyd.Field(
+        description="Non-hazardous waste disposed",
+        default=None,
+    )
+    rwd: ScopeSet | None = pyd.Field(
+        description="Radioactive waste disposed",
+        default=None,
+    )
+    hlrw: ScopeSet | None = pyd.Field(
+        description="High level radioactive waste disposed",
+        default=None,
+    )
+    illrw: ScopeSet | None = pyd.Field(
+        description="Intermediate level radioactive waste disposed",
+        default=None,
+    )
+    cru: ScopeSet | None = pyd.Field(
+        description="Components for re-use",
+        default=None,
+    )
+    mr: ScopeSet | None = pyd.Field(
+        description="Recycled materials",
+        default=None,
+    )
+    mfr: ScopeSet | None = pyd.Field(
+        description="Materials for recycling",
+        default=None,
+    )
+    mer: ScopeSet | None = pyd.Field(
+        description="Materials for energy recovery",
+        default=None,
+    )
+    ee: ScopeSet | None = pyd.Field(
+        description="Exported energy",
+        default=None,
+    )
+    eh: ScopeSet | None = pyd.Field(
+        description="Exported heat",
+        default=None,
+    )