+++ src/openepd/model/org.py
@@ -22,39 +22,36 @@ from typing import Annotated, Optional
 import pydantic as pyd
 
 from openepd.model.base import BaseOpenEpdSchema
-from openepd.model.common import WithAltIdsMixin, WithAttachmentsMixin
-
-
-class Contact(BaseOpenEpdSchema):  # TODO: NEW Object, not in the spec
-    """Contact information of a person or organization."""
-
-    email: pyd.EmailStr | None = pyd.Field(description="Email", example="contact@c-change-labs.com", default=None)
-    phone: str | None = pyd.Field(description="Phone number", example="+15263327352", default=None)
-    website: pyd.AnyUrl | None = pyd.Field(
-        description="Url of the website", example="http://buildingtransparency.org", default=None
-    )
-    address: str | None = pyd.Field(description="Address", example="123 Main St, San Francisco, CA 94105", default=None)
+from openepd.model.common import Location, WithAltIdsMixin, WithAttachmentsMixin
 
 
 class Org(WithAttachmentsMixin, WithAltIdsMixin, BaseOpenEpdSchema):
     """Represent an organization."""
 
-    web_domain: str = pyd.Field(
-        description="A web domain owned by organization. Typically is the org's home website address",
+    web_domain: str | None = pyd.Field(
+        description="A web domain owned by organization. Typically is the org's home website address", default=None
     )
-    name: str = pyd.Field(max_length=200, description="Common name for organization", example="C Change Labs")
+    name: str | None = pyd.Field(
+        max_length=200,
+        description="Common name for organization",
+        example="C Change Labs",
+        default=None,
+    )
     alt_names: Annotated[list[str], pyd.conlist(pyd.constr(max_length=200), max_items=255)] | None = pyd.Field(
         description="List of other names for organization",
         example=["C-Change Labs", "C-Change Labs inc."],
         default=None,
     )
     # TODO: NEW field, not in the spec
-    contacts: Contact | None = pyd.Field(description="Contact information of the organization", default=None)
     owner: Optional["Org"] = pyd.Field(description="Organization that controls this organization", default=None)
     subsidiaries: Annotated[list[str], pyd.conlist(pyd.constr(max_length=200), max_items=255)] | None = pyd.Field(
         description="Organizations controlled by this organization",
         example=["cqd.io", "supplychaincarbonpricing.org"],
         default=None,
+    )
+    hq_location: Location | None = pyd.Field(
+        default=None,
+        description="Location of a place of business, prefereably the corporate headquarters.",
     )
 
 
@@ -62,22 +59,29 @@ class Plant(WithAttachmentsMixin, WithAltIdsMixin, BaseOpenEpdSchema):
     """Represent a manufacturing plant."""
 
     # TODO: Add proper validator
-    id: str = pyd.Field(
+    id: str | None = pyd.Field(
         description="Plus code (aka Open Location Code) of plant's location and "
         "owner's web domain joined with `.`(dot).",
         example="865P2W3V+3W.interface.com",
         alias="pluscode",
+        default=None,
     )
     owner: Org | None = pyd.Field(description="Organization that owns the plant", default=None)
-    name: str = pyd.Field(
-        max_length=200, description="Manufacturer's name for plant. Recommended < 40 chars", example="Dalton, GA"
-    )
-    address: str = pyd.Field(
+    name: str | None = pyd.Field(
         max_length=200,
+        description="Manufacturer's name for plant. Recommended < 40 chars",
+        example="Dalton, GA",
+        default=None,
+    )
+    address: str | None = pyd.Field(
+        max_length=200,
+        default=None,
         description="Text address, preferably geocoded",
         example="1503 Orchard Hill Rd, LaGrange, GA 30240, United States",
     )
-    contact_email: pyd.EmailStr | None = pyd.Field(description="Email contact", example="info@interface.com")
+    contact_email: pyd.EmailStr | None = pyd.Field(
+        description="Email contact", example="info@interface.com", default=None
+    )
 
     class Config(BaseOpenEpdSchema.Config):
         allow_population_by_field_name = True